@@ -1,24 +1,24 @@
-import preprocess from 'svelte-preprocess';
 import adapter from '@sveltejs/adapter-static';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
+	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
+	// for more information about preprocessors
+	preprocess: vitePreprocess(),
+
 	kit: {
+		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
 		adapter: adapter({
-			// default options are shown
+			// default options are shown. On some platforms
+			// these options are set automatically — see below
 			pages: 'build',
 			assets: 'build',
-			fallback: null
-		}),
-		prerender: {
-			default: true
-		}
-	},
-	preprocess: [
-		preprocess({
-			postcss: true
+			fallback: undefined,
+			precompress: false,
+			strict: true
 		})
-	]
+	}
 };
 
 export default config;
